added business type model

+++ payfirstMerchant/src/main/java/com/payfirstMerchant/repository/TmpMerchantInfoRepository.java
@@ -3,5 +3,5 @@ package com.payfirstMerchant.repository;
 import com.payfirstMerchant.model.TmpMerchantInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TmpMerchantInfoRepository extends JpaRepository<TmpMerchantInfo, Long> {
+public interface TmpMerchantInfoRepository extends JpaRepository<TmpMerchantInfo, Integer> {
 }
